// src/adm/solicitudCompromiso/interfaces/puc/create.interfaces.ts
export interface CreatePuc {
    codigoPucSolicitud: number
    codigoDetalleSolicitud: number
    codigoSolicitud: number
    codigoSaldo: number
    codigoIcp: number
    codigoPuc: number
    financiadoId: number
    codigoFinanciado: number
    monto: number
    montoComprometido: number
    montoAnulado: number
    codigoPresupuesto: number
}